Put interfaces to different files
