+++ DynamicProperty/DynamicProperty.csproj
@@ -46,13 +46,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BasicValue.cs" />
-    <Compile Include="DependencyNode.cs" />
+    <Compile Include="IDependency.cs" />
     <Compile Include="DynamicValue.cs" />
     <Compile Include="IDependencySource.cs" />
     <Compile Include="CalculatedProperty.cs" />
     <Compile Include="BasicProperty.cs" />
     <Compile Include="DynamicProperty.cs" />
     <Compile Include="IDependencyTarget.cs" />
+    <Compile Include="IDependent.cs" />
     <Compile Include="IDynamicProperty{T}.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Subscription{T}.cs" />
