+++ DynamicProperty/IDependent.cs
@@ -1,0 +1,11 @@
+﻿using JetBrains.Annotations;
+
+namespace DynamicProperty
+{
+    interface IDependent
+    {
+        void DependsOn([NotNull] IDependency dependency);
+        void DoesNotDependOn([NotNull] IDependency dependency);
+        void Recalculate();
+    }
+}
